+++ Sources/Utils/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.37.0 (pre-release):
+* [Change] Don't reload the `KSPField` fields on language change to improve stability. Delegate it to the `Localization Tool`.
 
 # 0.36.0 (July 7th, 2018):
 * [Fix] Fix handling the negative values in all of the type formatters. E.g. [`GUIUtils.AngleType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_AngleType.htm).